requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.30
python-dotenv==1.0.0
requests==2.31.0
apscheduler==3.10.4
